ev3_robot/requirements.txt
evdev==1.4.0
paho-mqtt==1.5.1
PyYAML==5.4.1
pynput==1.7.3